% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_elo_params.R
\name{make_elo_params}
\alias{make_elo_params}
\title{make_elo_params}
\usage{
make_elo_params(
  home_advantage,
  k,
  mov_elo_add,
  mov_elo_mult,
  mov_score_add,
  mov_score_mult,
  score_factor
)
}
\arguments{
\item{score_factor}{}
}
\value{

}
\description{
This is just a handy way to set/change variables used by classes and functions. Will likely be removed
}
